Stricter strict mode

+++ myelin/kernel/generic-math.cc
@@ -28,7 +28,7 @@ namespace myelin {
 using namespace jit;
 
 static float sigmoid(float x) {
-  return 1.0 / (1.0 + exp(-x));
+  return 1.0 / (1.0 + expf(-x));
 }
 
 static float relu(float x) {
